Remove test_ffi and don't build mp4parse_capi as a dynamic library
These changes are necessary for conveniently vendoring into the
mozilla-central tree since test_ffi requires a dynamic library to avoid
duplicate symbols in the final binary, but certain builds fail in
mozilla-central if the vendored in crates build as dynamic libraries.

Since there is currently no way to parametrically disable building as a
dynamic library via cargo, we will remove test_ffi and rely on the builds
in the mozilla-central tree to alert us to API-breaking changes.